src/main/java/ru/practicum/user/UserController.java
package ru.practicum.user;


import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

import java.net.URLDecoder;
import java.net.URLEncoder;
import java.nio.charset.StandardCharsets;
import java.util.List;

@RestController
@RequiredArgsConstructor
@RequestMapping("/users")
public class UserController {
    private final UserService userService;

    @GetMapping
    public List<UserDto> getAllUsers() {
        return userService.getAllUsers();
    }

    @PostMapping
    public UserDto saveNewUser(@RequestBody UserDto userDto) {
        return userService.saveUser(userDto);
    }

    @GetMapping("/search")
    public List<UserShort> getShortUsers(@RequestParam String email) {
        //return userService.getShortUsers(URLDecoder.decode(email, StandardCharsets.UTF_8));
        return userService.getShortUsers(email);
    }
}
